refactor: move chackra ui provider to specific folter
